R6 Class for Pages Endpoint

Handle all pages operations in the Notion API

Note: Access this endpoint through the client instance, e.g., notion$pages. Not to be instantiated directly.

Value

A list containing the parsed API response.

Public fields

properties

Pages properties endpoint

Methods

Public methods

PagesEndpoint$new()

PagesEndpoint$create()

PagesEndpoint$retrieve()

Method new()

Initialise pages endpoint. Not to be called directly, e.g., use notion$pages instead.

Usage

PagesEndpoint$new(client)

Arguments

client

Notion Client instance

Method create()

Create a page

Usage

PagesEndpoint$create(
  parent,
  properties,
  children = NULL,
  icon = NULL,
  cover = NULL
)

Arguments

parent

Named list (JSON object) (required). The parent page or database where the new page is inserted.

properties

Named list (JSON object) (required). Key-value pairs representing the properties of the page.

children

List of lists (JSON array). Block objects to append as children to the page.

icon

Named list (JSON object). An icon for the page.

cover

Named list (JSON object). A cover image for the page.

Details

Endpoint documentation

Method retrieve()

Retrieve page properties

Usage

PagesEndpoint$retrieve(page_id, filter_properties = NULL)

Arguments

page_id

Character (required.). The ID for a Notion page.

filter_properties

Character. Page property value IDs to include in the response schema. If NULL, all properties are returned.

Details

Endpoint documentation

Examples

notion <- notion_client()
#> Error: Notion integration token not found.
#> ℹ Set it with `NOTION_TOKEN` environment variable (recommended)
#> ℹ or pass it directly with `auth` argument

# ----- create a page
notion$pages$create(
  list(page_id = "22f33ea0c1e480b99c77d1ab72aedff9"),
  list(
    title = list(list(
      text = list(
        content = "Test Page for notionapi"
      )
    ))
  )
)
#> Error: object 'notion' not found

# ----- retrieve a page
notion$pages$retrieve("23933ea0-c1e4-81d6-a6f6-dd5b57ad4aba")
#> Error: object 'notion' not found