Create empty property configuration

Helper function that creates an empty named list for property configurations that require no additional settings.

Many database properties like text, checkbox and date do not need configuration settings. This function returns the empty configuration ({} in JSON) that these properties expect.

Usage

no_config()

Value

An empty named list that serialises to {} in JSON

Examples

notion <- notion_client()

# ----- create a database
notion$databases$create(
  parent = list(page_id = "23933ea0-c1e4-81d6-a6f6-dd5b57ad4aba"),
  title = list(
    list(
      type = "text",
      text = list(
        content = "Grocery list"
      )
    )
  ),
  properties = list(
    Name = list(
      title = no_config()
    ),
    `In stock` = list(
      checkbox = no_config()
    )
  )
)
#> {
#>   "object": "database",
#>   "id": "23933ea0-c1e4-8136-b37b-fa235c6f2a71",
#>   "cover": {},
#>   "icon": {},
#>   "created_time": "2025-07-23T05:01:00.000Z",
#>   "created_by": {
#>     "object": "user",
#>     "id": "6b786605-e456-4237-9c61-5efaff23c081"
#>   },
#>   "last_edited_by": {
#>     "object": "user",
#>     "id": "6b786605-e456-4237-9c61-5efaff23c081"
#>   },
#>   "last_edited_time": "2025-07-23T05:01:00.000Z",
#>   "title": [
#>     {
#>       "type": "text",
#>       "text": {
#>         "content": "Grocery list",
#>         "link": {}
#>       },
#>       "annotations": {
#>         "bold": false,
#>         "italic": false,
#>         "strikethrough": false,
#>         "underline": false,
#>         "code": false,
#>         "color": "default"
#>       },
#>       "plain_text": "Grocery list",
#>       "href": {}
#>     }
#>   ],
#>   "description": [],
#>   "is_inline": false,
#>   "properties": {
#>     "In stock": {
#>       "id": "q%3BL%5E",
#>       "name": "In stock",
#>       "type": "checkbox",
#>       "checkbox": {}
#>     },
#>     "Name": {
#>       "id": "title",
#>       "name": "Name",
#>       "type": "title",
#>       "title": {}
#>     }
#>   },
#>   "parent": {
#>     "type": "page_id",
#>     "page_id": "23933ea0-c1e4-81d6-a6f6-dd5b57ad4aba"
#>   },
#>   "url": "https://www.notion.so/23933ea0c1e48136b37bfa235c6f2a71",
#>   "public_url": {},
#>   "archived": false,
#>   "in_trash": false,
#>   "request_id": "c3a4224b-5923-46c2-9497-c5527eb10b92"
#> }